requirements.txt
absl-py==2.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.60.0
fsspec==2024.6.1
grpcio==1.75.0
Jinja2==3.1.4
kiwisolver==1.4.9
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pillow==11.0.0
protobuf==6.32.1
psutil==7.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
